refactor cluster_fixture into multiple files
