+++ tests/cluster_fixture/kafka_config_templates/sasl_jaas.conf.j2
