+++ tests/cluster_fixture/kafka_config_templates/server.properties.j2
