+++ tests/cluster_fixture/kafka_config_templates/zookeeper.properties.j2
